@@ -2,6 +2,7 @@ plugins {
     id 'com.github.kt3k.coveralls' version '2.3.1'
     id 'de.aaschmid.cpd' version '1.0'
     id 'eclipse'
+    id 'idea'
     id 'groovy'
     id 'jacoco'
     id 'java'
@@ -14,7 +15,7 @@ description = 'A TestNG like dataprovider runner for JUnit having a simplified s
 group = 'com.tngtech.java'
 version = '1.12.1-SNAPSHOT'
 
-sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_6
+sourceCompatibility = targetCompatibility = JavaVersion.VERSION_1_8
 
 ext {
     isBuildOnJenkins = System.env['BUILD_TAG']?.startsWith('jenkins-') ?: false
@@ -43,11 +44,16 @@ configurations {
 }
 
 dependencies {
-    compile group: 'junit', name: 'junit', version: junitVersion
+//    compile group: 'junit', name: 'junit', version: junitVersion
 
     testCompile group: 'com.google.code.findbugs', name: 'annotations', version: '3.0.1'
     testCompile group: 'org.assertj', name: 'assertj-core', version: '1.7.1'
     testCompile group: 'org.mockito', name: 'mockito-core', version: '2.2.7'
+
+    // TODO: which can be runtime
+    compile group: 'org.junit.platform', name: 'junit-platform-runner', version: '1.0.0-M4'
+    compile group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: '5.0.0-M4'
+    compile group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.0.0-M4'
 
     integTestCompile group: 'org.codehaus.groovy', name: 'groovy', version: '2.4.7'
 }
@@ -154,6 +160,12 @@ artifacts {
     archives jar
     archives javadocJar
     archives sourcesJar
+}
+
+eclipse {
+    classpath {
+        plusConfigurations += [ configurations.integTestRuntime ]
+    }
 }
 
 // -- build and publish artifacts -------------------------------------------------------------------------------------
