@@ -16,8 +16,7 @@ public class SingleArgConverter extends AbstractObjectConverter<Object> {
      */
     @Override
     public Object[] convert(Object data, boolean isVarArgs, Class<?>[] parameterTypes) {
-        checkArgument(parameterTypes.length == 1, "Object[] dataprovider just supports single argument test method but found %d parameters",
-                parameterTypes.length);
+        checkArgument(parameterTypes.length == 1, "Object[] dataprovider just supports single argument test method but found %d parameters", parameterTypes.length);
         checkArgument(!isVarArgs, "Object[] dataprovider and single argumented test method does not support varargs");
 
         Object[] result = new Object[] { data };
